+++ x/posts/internal/types/msgs/msg_reactions_test.go
@@ -39,13 +39,13 @@ func TestMsgRegisterReaction_ValidateBasic(t *testing.T) {
 			name: "Empty short code returns error",
 			msg: msgs.NewMsgRegisterReaction(testOwner, "", "https://smile.jpg",
 				"4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e"),
-			error: sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "reaction short code must be an emoji short code"),
+			error: sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "The specified shortcode is not valid. To be valid it must only contains a-z, 0-9, - and _ and must start and end with a :"),
 		},
 		{
 			name: "Invalid short code returns error",
 			msg: msgs.NewMsgRegisterReaction(testOwner, ":smile", "https://smile.jpg",
 				"4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e"),
-			error: sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "reaction short code must be an emoji short code"),
+			error: sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "The specified shortcode is not valid. To be valid it must only contains a-z, 0-9, - and _ and must start and end with a :"),
 		},
 		{
 			name: "Empty value returns error",