+++ x/posts/internal/types/msgs/msgs_post_reactions_test.go
@@ -57,7 +57,7 @@ func TestMsgAddPostReaction_ValidateBasic(t *testing.T) {
 		{
 			name:  "Invalid value returns error",
 			msg:   msgs.NewMsgAddPostReaction(id, "like", testOwner),
-			error: sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "Reaction value must be an emoji or an emoji shortcode"),
+			error: sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "Reaction value must be an emoji or an emoji shortcode.If a shortcode is provided, it must only contains a-z, 0-9, - and _ and must start and end with a :"),
 		},
 		{
 			name:  "Valid message returns no error (with shortcode)",
@@ -128,9 +128,10 @@ func TestMsgRemovePostReaction_ValidateBasic(t *testing.T) {
 			error: sdkerrors.Wrap(sdkerrors.ErrInvalidAddress, "Invalid user address: "),
 		},
 		{
-			name:  "Blank value returns no error",
-			msg:   msgs.NewMsgRemovePostReaction(id, testOwner, ""),
-			error: sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "Reaction value must be an emoji or an emoji shortcode"),
+			name: "Blank value returns no error",
+			msg:  msgs.NewMsgRemovePostReaction(id, testOwner, ""),
+			error: sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "Reaction value must be an emoji or an emoji shortcode. "+
+				"If a shortcode is provided, it must only contains a-z, 0-9, - and _ and must start and end with a :"),
 		},
 		{
 			name:  "Valid message returns no error (with shortcode)",